change in vptree_factory? vptree_node

+++ lib/vptree_node.dart
@@ -1,7 +1,7 @@
 class VpTreeNode {
   int i;
-  int min;
-  int max;
+  double min;
+  double max;
   int mu;
   List<VpTreeNode> L;
   List<VpTreeNode> R;
